intel: only expose a 32bit ctx_id
todo: squash

+++ intel/intel_bufmgr_gem.c
@@ -3220,7 +3220,7 @@ drm_intel_gem_context_create(drm_intel_bufmgr *bufmgr)
 	return context;
 }
 
-drm_public uint64_t
+drm_public uint32_t
 drm_intel_gem_context_get_context_id(drm_intel_context *ctx)
 {
 	if (ctx == NULL)
